requirement.txt
fastapi
uvicorn
pydantic
PyJWT
python-decouple
pymongo